revision 1 for simulation

+++ src/citysimulation.c
@@ -20,9 +20,9 @@ CitySimulationResult* simulateCity(const City *city, int years, int *resultCount
     // Calculate the new population every year.
     double population = (double)city->population;
     for (int i = 0; i < years; i++) {
-        population = (size_t)(population * (1.0 + growthRate / 100.0));
+        population = population * (1.0 + growthRate / 100.0);
         results[i].year = i + 1;
-        results[i].population = population;
+        results[i].population = (size_t)population;
     }
 
     // Also set the user given result count pointer.
